+++ azure-storage-common/azure/storage/common/retry.py
@@ -159,7 +159,7 @@ class ExponentialRetry(_Retry):
     Exponential retry.
     '''
 
-    def __init__(self, initial_backoff=15, increment_power=3, max_attempts=3,
+    def __init__(self, initial_backoff=15, increment_base=3, max_attempts=3,
                  retry_to_secondary=False, random_jitter_range=3):
         '''
         Constructs an Exponential retry object. The initial_backoff is used for 
@@ -170,7 +170,7 @@ class ExponentialRetry(_Retry):
 
         :param int initial_backoff: 
             The initial backoff interval, in seconds, for the first retry.
-        :param int increment_power:
+        :param int increment_base:
             The base, in seconds, to increment the initial_backoff by after the 
             first retry.
         :param int max_attempts: 
@@ -184,7 +184,7 @@ class ExponentialRetry(_Retry):
             For example, a random_jitter_range of 3 results in the back-off interval x to vary between x+3 and x-3.
         '''
         self.initial_backoff = initial_backoff
-        self.increment_power = increment_power
+        self.increment_base = increment_base
         self.random_jitter_range = random_jitter_range
         super(ExponentialRetry, self).__init__(max_attempts, retry_to_secondary)
 
@@ -214,7 +214,7 @@ class ExponentialRetry(_Retry):
 
     def _backoff(self, context):
         random_generator = random.Random()
-        backoff = self.initial_backoff + (0 if context.count == 0 else pow(self.increment_power, context.count))
+        backoff = self.initial_backoff + (0 if context.count == 0 else pow(self.increment_base, context.count))
         random_range_start = backoff - self.random_jitter_range if backoff > self.random_jitter_range else 0
         random_range_end = backoff + self.random_jitter_range
         return random_generator.uniform(random_range_start, random_range_end)