+++ azure-storage-common/ChangeLog.md
@@ -5,6 +5,7 @@
 ## Version XX.XX.XX:
 - Added back the ability to generate account SAS for table service.
 - Fixed bug where a question mark prefix on SAS tokens causes failures.
+- The package has switched from Apache 2.0 to the MIT license.
 
 ## Version 0.37.1:
 - Fixed the return type of __add__ and __or__ methods on the AccountPermissions class